+++ e2e/tests/plugins.e2e.js
@@ -6,7 +6,7 @@ import {
   STATIC_TOOLBAR_BUTTONS,
   BUTTON_PLUGIN_MODAL,
 } from '../cypress/dataHooks';
-import { DEFAULT_DESKTOP_BROWSERS } from './settings';
+import { DEFAULT_DESKTOP_BROWSERS, DEFAULT_MOBILE_BROWSERS } from './settings';
 import { usePlugins, plugins, usePluginsConfig } from '../cypress/testAppConfig';
 
 const eyesOpen = ({
@@ -443,6 +443,38 @@ describe('plugins', () => {
 
     it('Change headers - without dropDownOptions config', () => {
       testHeaders(usePlugins(plugins.headings));
+    });
+  });
+
+  context('Text/Highlight Color - mobile', () => {
+    before(function() {
+      cy.eyesOpen({
+        appName: 'Text/Highlight Color - mobile',
+        testName: this.test.parent.title,
+        browser: DEFAULT_MOBILE_BROWSERS,
+      });
+    });
+    beforeEach(() => cy.switchToMobile());
+
+    after(() => cy.eyesClose());
+
+    it('allow to color text', function() {
+      cy.loadRicosEditorAndViewer()
+        .enterParagraphs(['Color.'])
+        .setTextColor([0, 5], 'color4');
+
+      cy.blurEditor();
+
+      cy.eyesCheckWindow(this.test.title);
+    });
+
+    it('allow to highlight text', function() {
+      cy.loadRicosEditorAndViewer()
+        .enterParagraphs(['Highlight.'])
+        .setHighlightColor([0, 9], 'color4');
+
+      cy.blurEditor();
+      cy.eyesCheckWindow(this.test.title);
     });
   });
 });