+++ CHANGELOG.md
@@ -15,8 +15,9 @@
     Click to see more.
   </summary>
 
-
-
+### :bug: Bug Fix
+- `text-color`
+  - [#1279](https://github.com/wix-incubator/rich-content/pull/1279) text & highlight color in mobile
 </details>
 <hr/>
 
@@ -27,7 +28,7 @@
 - `gallery`
   - [#1273](https://github.com/wix-incubator/rich-content/pull/1273) gallery opens on correct image in rtl
 - `mentions`
-- - [#1275](https://github.com/wix-incubator/rich-content/pull/1275) `onMentionClick` callback is called on viewer
+  - [#1275](https://github.com/wix-incubator/rich-content/pull/1275) `onMentionClick` callback is called on viewer
 - `common`
   - [#1274](https://github.com/wix-incubator/rich-content/pull/1274) viewer text direction
 ### :house: Internal